Erstatte gammel komponent med ny
